+++ src/download/message.cpp
@@ -3,24 +3,8 @@
 #include <array>
 
 namespace fur::download::message {
-/// Takes a 32 bits unsigned integer and encodes it into a 4 bytes array.
-std::array<uint8_t, 4> encode_big_endian(uint32_t n) {
-  std::array<uint8_t, 4> result{};
-  for (int i = 3; i >= 0; i--) {
-    result[i] = n & 0xFF;
-    n >>= 8;
-  }
-  return result;
-}
+#include "download/util.hpp"
 
-/// Decodes a 32 bits unsigned integer from a 4 bytes array.
-uint32_t decode_big_endian(const std::array<uint8_t, 4>& buf) {
-  uint32_t result = 0;
-  for (int i = 0; i <= 3; i++) {
-    result |= buf[i] << (8 * (3 - i));
-  }
-  return result;
-}
 
 std::vector<uint8_t> Message::encode() {
   // Collect bytes here
